move: user repository 디렉터리 파일 분류
