+++ src/main/java/com/newzet/api/user/repository/repository/UserJpaRepository.java
@@ -1,9 +1,11 @@
-package com.newzet.api.user.repository;
+package com.newzet.api.user.repository.repository;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
+import com.newzet.api.user.repository.entity.UserEntity;
 
 @Repository
 public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
